Subpixel rendering: Use floorf/roundf/fabs in device snapping helpers. https://bugs.webkit.org/show_bug.cgi?id=128075
Reviewed by Darin Adler.

No change in functionality.

* platform/LayoutUnit.h:
(WebCore::roundToDevicePixel):
(WebCore::floorToDevicePixel):
* platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContext::platformInit):


Canonical link: https://commits.webkit.org/146096@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163278 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/LayoutUnit.h
@@ -937,12 +937,12 @@ inline int floorToInt(LayoutUnit value)
 
 inline float roundToDevicePixel(LayoutUnit value, float pixelSnappingFactor)
 {
-    return round((value.rawValue() * pixelSnappingFactor) / kEffectiveFixedPointDenominator) / pixelSnappingFactor;
+    return roundf((value.rawValue() * pixelSnappingFactor) / kEffectiveFixedPointDenominator) / pixelSnappingFactor;
 }
 
 inline float floorToDevicePixel(LayoutUnit value, float pixelSnappingFactor)
 {
-    return floor((value.rawValue() * pixelSnappingFactor) / kEffectiveFixedPointDenominator) / pixelSnappingFactor;
+    return floorf((value.rawValue() * pixelSnappingFactor) / kEffectiveFixedPointDenominator) / pixelSnappingFactor;
 }
 
 inline float snapSizeToDevicePixel(LayoutUnit size, LayoutUnit location, float pixelSnappingFactor)
